+++ app/src/main/java/woowacourse/shopping/data/product/ProductRemoteRepository.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.data.product
 
 import com.example.domain.Pagination
-import com.example.domain.Product
-import com.example.domain.repository.ProductRepository
+import com.example.domain.product.Product
+import com.example.domain.product.ProductRepository
 import retrofit2.Call
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
