+++ app/src/main/java/woowacourse/shopping/data/product/RetrofitProductService.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.data.product
 
-import com.example.domain.Product
+import com.example.domain.product.Product
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Path
