+++ app/src/main/java/woowacourse/shopping/feature/product/MainActivity.kt
@@ -8,12 +8,12 @@ import android.view.View.VISIBLE
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.ConcatAdapter
 import androidx.recyclerview.widget.GridLayoutManager
-import com.example.domain.CartProduct
-import com.example.domain.Product
-import com.example.domain.RecentProduct
-import com.example.domain.repository.CartRepository
-import com.example.domain.repository.ProductRepository
-import com.example.domain.repository.RecentProductRepository
+import com.example.domain.cart.CartProduct
+import com.example.domain.cart.CartRepository
+import com.example.domain.product.Product
+import com.example.domain.product.ProductRepository
+import com.example.domain.product.recent.RecentProduct
+import com.example.domain.product.recent.RecentProductRepository
 import woowacourse.shopping.ServerType
 import woowacourse.shopping.common.adapter.LoadMoreAdapter
 import woowacourse.shopping.data.cart.CartRemoteRepository
