+++ app/src/main/java/woowacourse/shopping/feature/product/MainContract.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.feature.product
 
-import com.example.domain.CartProduct
-import com.example.domain.Product
-import com.example.domain.RecentProduct
+import com.example.domain.cart.CartProduct
+import com.example.domain.product.Product
+import com.example.domain.product.recent.RecentProduct
 import woowacourse.shopping.model.CartProductState
 import woowacourse.shopping.model.ProductState
 import woowacourse.shopping.model.RecentProductState
