+++ app/src/main/java/woowacourse/shopping/feature/product/MainPresenter.kt
@@ -1,12 +1,12 @@
 package woowacourse.shopping.feature.product
 
-import com.example.domain.Cart
-import com.example.domain.CartProduct
 import com.example.domain.Pagination
-import com.example.domain.Product
-import com.example.domain.repository.CartRepository
-import com.example.domain.repository.ProductRepository
-import com.example.domain.repository.RecentProductRepository
+import com.example.domain.cart.Cart
+import com.example.domain.cart.CartProduct
+import com.example.domain.cart.CartRepository
+import com.example.domain.product.Product
+import com.example.domain.product.ProductRepository
+import com.example.domain.product.recent.RecentProductRepository
 import woowacourse.shopping.model.CartProductState
 import woowacourse.shopping.model.CartProductState.Companion.MAX_COUNT_VALUE
 import woowacourse.shopping.model.CartProductState.Companion.MIN_COUNT_VALUE
