+++ app/src/main/java/woowacourse/shopping/model/mapper/CartProductMapper.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.model.mapper
 
-import com.example.domain.CartProduct
+import com.example.domain.cart.CartProduct
 import woowacourse.shopping.model.CartProductState
 
 fun CartProduct.toUi(): CartProductState {
