+++ app/src/main/java/woowacourse/shopping/model/mapper/RecentProductMapper.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.model.mapper
 
-import com.example.domain.Product
-import com.example.domain.RecentProduct
+import com.example.domain.product.Product
+import com.example.domain.product.recent.RecentProduct
 import woowacourse.shopping.model.RecentProductState
 
 fun RecentProduct.toUi(): RecentProductState {
