+++ domain/src/main/java/com/example/domain/cart/CartProduct.kt
@@ -1,4 +1,4 @@
-package com.example.domain
+package com.example.domain.cart
 
 class CartProduct(
     val id: Long,
