+++ domain/src/main/java/com/example/domain/order/OrderRepository.kt
@@ -1,8 +1,6 @@
-package com.example.domain.repository
+package com.example.domain.order
 
 import com.example.domain.FixedDiscountPolicy
-import com.example.domain.order.Order
-import com.example.domain.order.OrderSummary
 
 interface OrderRepository {
