+++ domain/src/main/java/com/example/domain/product/Product.kt
@@ -1,4 +1,4 @@
-package com.example.domain
+package com.example.domain.product
 
 data class Product(
     val id: Long,
