+++ domain/src/main/java/com/example/domain/product/recent/RecentProduct.kt
@@ -1,4 +1,4 @@
-package com.example.domain
+package com.example.domain.product.recent
 
 import java.time.LocalDateTime
